Add enumerate() examples to supplies_list_range_ex.py
Although using `for i in range(len(supplies))` works,
you will still need to look up the index of each item
using `i` (i.e., `supplies[i]`). This is not the most
efficient approach.

Instead, consider using `for i,item in enumerate(supplies)`.
This will return an index (`i`) and the item corresponding
to that index within supplies (i.e., `item` == `supplies[i]`).
With this approach, enumerate() takes care of the indexing
for you and is more efficient.

Resolves: #4