Added examples for %-formatting, str.format(), and f-strings
Python offers a wide variety of how to print objects.
The best approach is to use f-strings as they are fast
and have some great attributes. Since f-strings are available
in Python 3.6 and above, you may have to revert to using
str.format() in some cases. It is recommended to not use
%-formatting anymore.

Closes: #7